+++ src/test/java/org/mmarini/railways2/model/trains/TrainRunningAtPlatformTest.java
@@ -35,8 +35,8 @@ import org.mmarini.railways2.model.StationStatus;
 import org.mmarini.railways2.model.geometry.*;
 import org.mmarini.railways2.model.route.Entry;
 import org.mmarini.railways2.model.route.Exit;
+import org.mmarini.railways2.model.route.Junction;
 import org.mmarini.railways2.model.route.RoutesConfig;
-import org.mmarini.railways2.model.route.Signal;
 
 import java.awt.geom.Point2D;
 import java.util.List;
@@ -44,18 +44,19 @@ import java.util.Map;
 import java.util.Optional;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mmarini.railways.TestFunctions.locatedAt;
-import static org.mmarini.railways2.model.RailwayConstants.LOADING_TIME;
+import static org.mmarini.railways2.model.RailwayConstants.MAX_SPEED;
 
-class TrainWaitForLoadedTest {
+class TrainRunningAtPlatformTest {
     static final double DT = 0.1;
     static RoutesConfig routes;
     static StationMap stationMap;
 
     /**
      * <pre>
-     *     Entry(a) --ab(500m)-- Signal(b) --bc(500m)-- Exit(c)
+     *     Entry(a) --ab(500m)-- Junction(b) --bcPlatform(500m)-- Junction(c) --cd(500m).. Exit(d)
      * </pre>
      */
     @BeforeAll
@@ -63,55 +64,59 @@ class TrainWaitForLoadedTest {
         stationMap = new StationBuilder("station")
                 .addNode("a", new Point2D.Double(), "ab")
                 .addNode("b", new Point2D.Double(500, 0), "ab", "bc")
-                .addNode("c", new Point2D.Double(1000, 0), "bc")
+                .addNode("c", new Point2D.Double(1000, 0), "bc", "cd")
+                .addNode("d", new Point2D.Double(1500, 0), "cd")
                 .addEdge(Track.builder("ab"), "a", "b")
-                .addEdge(Track.builder("bc"), "b", "c")
+                .addEdge(Platform.builder("bc"), "b", "c")
+                .addEdge(Track.builder("cd"), "c", "d")
                 .build();
         routes = RoutesConfig.create(stationMap, Map.of(
                 "a", Entry::new,
-                "b", Signal::create,
-                "c", Exit::new
+                "b", Junction::new,
+                "c", Junction::new,
+                "d", Exit::new
         ));
     }
 
     @Test
-    void loaded() {
-        Entry arrival = routes.getRoute("a");
-        Exit destination = routes.getRoute("c");
-        Edge ab = stationMap.getEdge("ab");
-        Train train = Train.create("train2", 1, arrival, destination)
-                .setLocation(new OrientedLocation(ab, true, ab.getLength()))
-                .load(0);
-        StationStatus status = StationStatus.create(stationMap, routes, List.of(train))
-                .setTime(LOADING_TIME);
+    void runningPlatform() {
+        Entry a = routes.getRoute("a");
+        Exit d = routes.getRoute("d");
+        Edge bc = stationMap.getEdge("bc");
+        Edge cd = stationMap.getEdge("cd");
+        double distance = 497; // edge length=500m, moving distance=3.6m, distance = 500-3 = 497m
+        Train train = Train.create("train2", 1, a, d)
+                .setLocation(new OrientedLocation(bc, true, distance))
+                .setLoaded(true)
+                .setState(Train.State.RUNNING_TRAIN_STATE);
+        StationStatus status = StationStatus.create(stationMap, routes, List.of(train));
 
-        Optional<Train> nextOpt = train.loading(new SimulationContext(status, DT));
+        Optional<Train> nextOpt = train.running(new SimulationContext(status, DT));
 
         assertTrue(nextOpt.isPresent());
         Train next = nextOpt.orElseThrow();
-        assertEquals(Train.State.WAITING_FOR_RUN_TRAIN_STATE, next.getState());
-        assertEquals(0, next.getSpeed());
-        assertThat(next.getLocation(), locatedAt(ab, true, ab.getLength()));
-        assertTrue(next.isLoaded());
+        assertEquals(Train.State.RUNNING_TRAIN_STATE, next.getState());
+        assertEquals(MAX_SPEED, next.getSpeed());
+        assertThat(next.getLocation(), locatedAt(cd, true, distance - bc.getLength() + DT * MAX_SPEED));
     }
 
     @Test
-    void waitForLoad() {
-        Entry arrival = routes.getRoute("a");
-        Exit destination = routes.getRoute("c");
-        Edge ab = stationMap.getEdge("ab");
-        Train train = Train.create("train2", 1, arrival, destination)
-                .setLocation(new OrientedLocation(ab, true, ab.getLength()))
-                .load(0);
+    void stoppingPlatform() {
+        Entry a = routes.getRoute("a");
+        Exit d = routes.getRoute("d");
+        Edge bc = stationMap.getEdge("bc");
+        double distance = 497; // edge length=500m, moving distance=3.6m, distance = 500-3 = 497m
+        Train train = Train.create("train2", 1, a, d)
+                .setLocation(new OrientedLocation(bc, true, distance))
+                .setState(Train.State.RUNNING_TRAIN_STATE);
         StationStatus status = StationStatus.create(stationMap, routes, List.of(train));
 
-        Optional<Train> nextOpt = train.loading(new SimulationContext(status, DT));
+        Optional<Train> nextOpt = train.running(new SimulationContext(status, DT));
 
         assertTrue(nextOpt.isPresent());
         Train next = nextOpt.orElseThrow();
         assertEquals(Train.State.LOADING_TRAIN_STATE, next.getState());
         assertEquals(0, next.getSpeed());
-        assertThat(next.getLocation(), locatedAt(ab, true, ab.getLength()));
-        assertFalse(next.isLoaded());
+        assertThat(next.getLocation(), locatedAt(bc, true, bc.getLength()));
     }
 }